+++ storyblok/RichtextYoutube.vue
@@ -4,5 +4,5 @@ defineProps({ blok: Object });
 
 <template>
   <div v-if="!blok.video_id">Please provide a YouTube video ID.</div>
-  <iframe v-else :src="`https://www.youtube-nocookie.com/embed/${blok.video_id}`" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope;" allowfullscreen class="aspect-video w-full"></iframe>
+  <iframe v-else :src="`https://www.youtube-nocookie.com/embed/${blok.video_id}`" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope;" allowfullscreen class="aspect-video w-full rounded-lg"></iframe>
 </template>